Backed out 9 changesets (bug 1596845) for causing xpcshell failures on test_trr_enterprise_policy.js. CLOSED TREE
Backed out changeset 231acfc052bb (bug 1596845)
Backed out changeset 7d02dad4d720 (bug 1596845)
Backed out changeset b0449eec2671 (bug 1596845)
Backed out changeset 5cd033c9ef7c (bug 1596845)
Backed out changeset 98d583f1d19e (bug 1596845)
Backed out changeset b78bc0049605 (bug 1596845)
Backed out changeset 88919aaff89b (bug 1596845)
Backed out changeset f4adb3e7b8e1 (bug 1596845)
Backed out changeset b171f7ac0b40 (bug 1596845)

+++ toolkit/components/bitsdownload/src/bits_interface/request.rs
@@ -713,16 +713,16 @@ impl BitsRequest {
     }
 
     xpcom_method!(
-        get_trr_mode => GetTRRMode() -> u32
+        get_trr_mode => GetTRRMode() -> u8
     );
-    fn get_trr_mode(&self) -> Result<u32, nsresult> {
+    fn get_trr_mode(&self) -> Result<u8, nsresult> {
         Err(NS_ERROR_NOT_IMPLEMENTED)
     }
 
     xpcom_method!(
-        set_trr_mode => SetTRRMode(_trr_mode: u32)
+        set_trr_mode => SetTRRMode(_trr_mode: u8)
     );
-    fn set_trr_mode(&self, _trr_mode: u32) -> Result<(), nsresult> {
+    fn set_trr_mode(&self, _trr_mode: u8) -> Result<(), nsresult> {
         Err(NS_ERROR_NOT_IMPLEMENTED)
     }
